requirements.txt
django==2.0.5
django-extensions==2.0.7
django-grappelli==2.11.1
django-model-utils==3.1.2
django-allauth==0.36.0
django-environ==0.4.4
